requirements.txt
django>=5.0,<6.0
psycopg[binary]>=3.1
pgvector>=0.3.2
djangorestframework>=3.15
python-dotenv>=1.0
requests>=2.32
sentence-transformers>=3.0
httpx>=0.27
trafilatura>=1.9
pypdf>=4.0
beautifulsoup4>=4.12
